Add justfile, fix all clippy warnings, clean up tests
justfile with recipes for:
- test/test-unit/test-integration/test-one (sudo for CAP_NET_RAW)
- example shortcuts (capture, batch, inject, fanout, stats, etc.)
- ci/ci-full (lint + test + docs + bench)
- check-afpacket (detect container limitations)
- clippy, fmt, doc, bench, clean, project-stats

Clippy fixes:
- Remove redundant u32/usize casts (TPACKET_ALIGNMENT/TPACKET3_HDRLEN
  already usize in libc, PACKET_FANOUT_* already u32)
- Use .div_ceil() instead of manual (a+b-1)/b
- Use io::Error::other() instead of Error::new(Other, _)
- Derive Default instead of manual impl for InjectorBuilder
- Fix length-comparison-to-zero in tests (use !is_empty())
- Allow dead_code in test helpers module
- Remove unused import in fanout test
- Fix unnecessary mut in statistics test

Zero clippy errors. CI passes. 70 tests.

+++ src/afpacket/ffi.rs
@@ -56,7 +56,7 @@ pub use libc::TPACKET_ALIGNMENT;
 
 /// Align `x` up to `TPACKET_ALIGNMENT` (16 bytes).
 pub const fn tpacket_align(x: usize) -> usize {
-    (x + (TPACKET_ALIGNMENT as usize) - 1) & !((TPACKET_ALIGNMENT as usize) - 1)
+    (x + TPACKET_ALIGNMENT - 1) & !(TPACKET_ALIGNMENT - 1)
 }
 
 // ── Feature request flags ──────────────────────────────────────────────────
